[JetBrains] add more integration tests
